Cleanup and add tests
