+++ edge-agent/src/Microsoft.Azure.Devices.Edge.Agent.IoTHub/blob/AzureBlob.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 namespace Microsoft.Azure.Devices.Edge.Agent.IoTHub.Blob
 {
-    using System;
-    using System.IO;
     using System.Threading.Tasks;
     using Microsoft.Azure.Devices.Edge.Util;
     using Microsoft.WindowsAzure.Storage.Blob;
@@ -19,19 +17,5 @@ namespace Microsoft.Azure.Devices.Edge.Agent.IoTHub.Blob
         public string Name => this.blockBlob.Name;
 
         public Task UploadFromByteArrayAsync(byte[] bytes) => this.blockBlob.UploadFromByteArrayAsync(bytes, 0, bytes.Length);
-    }
-
-    class AzureAppendBlob : IAzureAppendBlob
-    {
-        readonly CloudAppendBlob appendBlob;
-
-        public AzureAppendBlob(CloudAppendBlob appendBlob)
-        {
-            this.appendBlob = Preconditions.CheckNotNull(appendBlob, nameof(appendBlob));
-        }
-
-        public string Name => this.appendBlob.Name;
-
-        public Task AppendByteArray(ArraySegment<byte> bytes) => this.appendBlob.AppendFromByteArrayAsync(bytes.Array, bytes.Offset, bytes.Count);
     }
 }
